third_party/WebKit/LayoutTests/http/tests/navigation/same-document-scroll-position-restore-pushstate.html

<!DOCTYPE html>
<body>
<script src="/js-test-resources/js-test.js"></script>
<div style="height: 1100px">scroll down...</div>
<div id="log"></div>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
}

window.onload = function() {
    window.scrollBy(0, 1000);
    history.pushState({}, '', '');
    history.pushState({}, '', '');
    scrollOffsetY = window.scrollY;
    history.back();
    setTimeout(function() {
        shouldBeTrue('window.scrollY == scrollOffsetY');
        if (window.testRunner)
            testRunner.notifyDone();
    }, 0);
};
</script>
</body>
