third_party/WebKit/LayoutTests/http/tests/navigation/same-document-scroll-position-restore-pushstate-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS window.scrollY == scrollOffsetY is true
scroll down...
